+++ core-bundle/tests/Cors/WebsiteRootsConfigProviderTest.php
@@ -14,7 +14,7 @@ use Contao\CoreBundle\Cors\WebsiteRootsConfigProvider;
 use Doctrine\DBAL\Connection;
 use Doctrine\DBAL\Driver\Mysqli\MysqliException;
 use Doctrine\DBAL\Driver\Statement;
-use Doctrine\DBAL\Exception\ConnectionException;
+use Doctrine\DBAL\Exception\DriverException;
 use Doctrine\DBAL\Schema\MySqlSchemaManager;
 use PHPUnit\Framework\TestCase;
 use Symfony\Component\HttpFoundation\Request;
@@ -152,7 +152,7 @@ class WebsiteRootsConfigProviderTest extends TestCase
 
         $connection
             ->method('isConnected')
-            ->willThrowException(new ConnectionException('Could not connect', new MysqliException('Invalid password')))
+            ->willThrowException(new DriverException('Could not connect', new MysqliException('Invalid password')))
         ;
 
         $connection